requirements.txt
# FastAPI e server
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0

# PaddlePaddle CPU - versione flessibile
paddlepaddle>=2.6.0,<3.0.0

# PaddleOCR - versione flessibile
paddleocr>=2.7.0,<3.0.0

# Image processing - versioni flessibili per ARM64
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=9.5.0,<11.0.0
numpy>=1.24.0,<2.0.0

# Scientific computing per rotation detection avanzata
scipy>=1.10.0,<2.0.0

# Utilities base
requests>=2.31.0,<3.0.0

pytesseract>=0.3.13  # Versione più recente che dovrebbe supportare NumPy 2.x

# Document processing - basic
PyPDF2>=3.0.0
python-pptx>=0.6.21
python-docx>=0.8.11
openpyxl>=3.0.10

# Document processing - advanced
#unstructured[pdf]>=0.15.0  # Aggiunto supporto PDF e versione più recente
PyMuPDF>=1.24.0  # Versione più recente
pdfplumber>=0.11.0  # Versione più recente
